created a new file for menu actions and notes_view
