+++ lib/services/auth/auth_service.dart
@@ -1,10 +1,13 @@
 import 'auth_provider.dart';
 import 'auth_user.dart';
+import 'firebase_auth_provider.dart';
 
 class AuthService implements AuthProvider {
   final AuthProvider provider;
 
   const AuthService(this.provider);
+
+  factory AuthService.firebase() => AuthService(FirebaseAuthProvider());
 
   @override
   Future<AuthUser> createUser({
